Enable neovim system clipboard support in Linux
Add `unnamedplus` to enable clipboard support in Neovim in Linux, as
described in [this StackOverflow post].

[this StackOverflow post]: https://stackoverflow.com/questions/30691466/what-is-difference-between-vims-clipboard-unnamed-and-unnamedplus-settings

This requires the installation of `xclip` over `xsel` in order to work.
Also, the extra installation of `autocutsel` and the extra lines in
`.xinitrc` allow for text from a Neovim buffer to remain in the X11
clipboard even after Neovim has been closed. These changes replicate the
expected clipboard behavior of Windows and macOS.