Fix glob sourcing of bash completions on macOS
It turns out that passing a globbed path to `source` in Bash doesn't
actually work. The command expects the first argument to be a file path,
and the rest are arguments passed to the file being sourced. As such,
the correct usage would be to `for` loop through each file and `source`
it individually, which with this commit the `.bashrc` now does.