Install all Neovim plugins before setting options
It turns out that certain settings that are called before plugins are
loaded, e.g. `syntax on`, will take effect using the old, built-in
versions of the plugins, e.g. `rust.vim`, meaning that the newer
versions from `vim-plug` never get loaded, or get partially loaded.

This commit moves the `plug#begin()` and `plug#end()` calls all the way
to the beginning of the `init.vim` to prevent this from happening.