Add puremourning/vimspector plugin to Neovim
This plugin lets us debug C/C++/Rust applications using GDB or LLDB. It
is roughly equivalent to this VSCode plugin:

https://marketplace.visualstudio.com/items?itemName=vadimcn.vscode-lldb

Note that `vimspector` doesn't have 100% compatibility with Neovim,
particularly the WinBar. In this config, we include a several extra key
mappings which replicate the buttons which would be visible in the
WinBar, were we using Vim instead of Neovim.