Improve terminal-overrides defined in .tmux.conf
This commit switches to "tmux-256color" for the italic text support and
adds support for [undercurls] and separately-colored underlines. This is
in preparation for a future Neovim configuration revamp.

[undercurls]: https://github.com/folke/lsp-colors.nvim/blob/4e6da1984d23da88a947805866580c48fc3cc8d7/README.md#making-undercurls-work-properly-in-tmux

We also add an `:RGB` terminal override to please Neovim. Uncovered by
running `:checkhealth`, this ensures proper 256 color support while
Neovim is nested within a `tmux` session.