Add nvim-telescope/telescope-ui-select.nvim plugin
This commit adds support for the `ui-select` extension to Telescope in
Neovim, which overrides the default `vim.ui.select()` handler and
displays a nice-looking dropdown dialog instead.

Because Telescope itself is being lazy-loaded on editor startup (it
consistently slows down editor startup according to `vim-startuptime`),
this `ui-select` extension would not work at first. As such, we provide
a custom `setup` function which tells Telescope to load itself, if not
already loaded, when `vim.ui.select()` is called for the first time.