Fix Jinja2 file detection in Neovim
There's a chicken-and-egg problem here: we can't lazy-load the plugin
using `ft = "jinja"` if the plugin hasn't already been loaded yet.
Better to suck it up and always load the plugin and do so eagerly. On
the bright side, it seems to be a pretty small and fast-loading plugin.
It doesn't seem to affect the readings in `vim-startuptime` very much
(barely 0.035200 milliseconds on average, at the time of writing).