Refactor 'top' package, use --stow during install
This commit refactors the `top` package to be more intuitive: the
`config-linux` package contains my personal config, while `config-work`
contains my work config.

The `.stowrc` ensures that the `toprc` from the "work" profile overrides
the base `toprc` upon installation, and the `config_precleanup_linux()`
hook ensures that `dotfiles uninstall` works as expected.

Key to making this work is the switch to using `--stow` during
installation instead of `--restow`. If this change isn't done, `stow`
will complain if you try to do:

1. `dotfiles install top`
2. `dotfiles install top+work` (error: conflict!)

This may have negative implications for other packages, though nothing
seems to be behaving out of the ordinary when testing in a sandbox. I'm
not entirely sure.