Make 'dotfiles' command reachable from $PATH
The approach taken in this commit is broken down as follows:

1. The `bash` stow package now designates ~/.local/bin as a special
   directory for user-specific executables. This adheres to the
   convention defined by the XDG Base Directory spec (refer to the
   "Environment Variables" section):

   https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html

   We make sure this same directory works on macOS, which doesn't
   have an equivalent convention of its own for non-GUI apps.

2. The `bash` stow package now includes a ~/.local/bin/dotfiles symlink
   which points to the `dotfiles` script at the repo root. This makes
   the `dotfiles` command automatically accessible from anywhere via
   `$PATH` if users install the `bash` stow package.

That takes care of setting up our `$PATH`, but for the `dotfiles` script
to actually work from any directory, it needs to know where to find all
the stow packages.

To solve this, we update the `dotfiles` script to locate its own path
using `dirname $(readlink -f ${BASH_SOURCE[0]})` and look for stow
packages there.

NOTE: It seems that `readlink -f` was silently introduced in macOS
Monterey 12.3 (March 2022) and `realpath` in macOS Ventura 13.0 (October
2022), so this should as expected across both Linux and macOS.