Add hacky-but-working fzf completion for 'git checkout **'
This is incredibly cursed, but I'm impressed it works so well.

Note that the weird clause messing with `$COMPREPLY[-1]` there is
trimming the trailing whitespace from the last element. This ensures
that we don't complete e.g. `git checko<|>` as `git checkout  ` (note
the double space at the end). By trimming this extra space at the end,
FZF is able to produce the completion `git checkout ` (one trailing
space), which is the expected result.