Add 'mopidy' as opt-in alternative to 'mpd'
The primary advantage of Mopidy over MPD for me is support for streaming
music, e.g. Jellyfin or Spotify or YouTube Music.

This commit includes a Linux configuration (as a systemd user service,
for compatibility with wireplumber) and a macOS configuration, similar
to the existing `mpd` package. Note that the macOS config is provided on
a BEST EFFORT basis since I don't currently have a Macbook on hand to
test it with.

There are some caveats surrounding the `ncmpcpp` music visualizer. These
caveats are as follows:

1. Mopidy cannot properly produce a `/tmp/mpd.fifo` pipe which the
   `ncmpcpp` visualizer expects from MPD. I've tried `filesink
   location=/tmp/mpd.fifo` in `mopidy.conf`, and I see major stuttering
   in the `ncmpcpp` visualizer.

2. However, Mopidy _does_ support serving visualizer data at
   `localhost:5555` and `ncmpcpp` supports this too, and it works
   beautifully. Unfortunately, MPD doesn't support this mode, meaning
   "ncmpcpp w/MPD" and "ncmpcpp w/Mopidy" can't share the same `ncmpcpp`
   configuration.

As such, this commit also updates the install semantics of the `ncmpcpp`
package. If you `dotfiles install ncmpcpp`, the configs are set up for
MPD. If you `dotfiles install ncmpcpp+mopidy`, the configs are set up
for Mopidy.

If I can verify that Mopidy works just as well as MPD on macOS, I might
just ditch MPD permanently and remove this `ncmpcpp` "alt profile" hack.