Fix Debian/Ubuntu fzf bash-completion
It seems that Debian/Ubuntu broke the `bash-completion` integration for
the `fzf` package at some point, as per the ticket linked below:

https://github.com/junegunn/fzf/issues/2790

Those breaking changes were seemingly reverted in `fzf >= 0.60.0-1`.
Since my work computer is currently stuck on an older version of `fzf`,
this commit includes a temporary workaround for this issue.